Fixed warnings on Win64

+++ src/dist.cpp
@@ -139,7 +139,7 @@ void zmq::dist_t::distribute (msg_t *msg_, int flags_)
 
     //  Add matching-1 references to the message. We already hold one reference,
     //  that's why -1.
-    msg_->add_refs (matching - 1);
+    msg_->add_refs ((int) matching - 1);
 
     //  Push copy of the message to each matching pipe.
     for (pipes_t::size_type i = 0; i < matching;) {
